+++ packages/controlled-form/src/lib/TextField.tsx
@@ -36,6 +36,13 @@ export const ControlledTextField = ({
           required={typeof rules.required === 'object' ? rules.required.value : !!rules.required}
           {...field}
           {...rest}
+          slotProps={{
+            ...rest.slotProps,
+            htmlInput: {
+              maxLength: typeof rules.maxLength === 'object' ? rules.maxLength.value : rules.maxLength,
+              ...rest.slotProps?.htmlInput,
+            },
+          }}
           inputRef={ref}
           error={!!error}
           helperText={