+++ packages/autocomplete/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.0.8](https://github.com/Availity/element/compare/@availity/mui-autocomplete@1.0.7...@availity/mui-autocomplete@1.0.8) (2025-03-27)
+
+### Dependency Updates
+
+* `mui-textfield` updated to version `1.0.7`
+## [1.0.7](https://github.com/Availity/element/compare/@availity/mui-autocomplete@1.0.6...@availity/mui-autocomplete@1.0.7) (2025-03-21)
+
+### Dependency Updates
+
+* `mui-form-utils` updated to version `1.0.6`
+* `mui-textfield` updated to version `1.0.6`
 ## [1.0.6](https://github.com/Availity/element/compare/@availity/mui-autocomplete@1.0.5...@availity/mui-autocomplete@1.0.6) (2025-03-13)
 
 ### Dependency Updates