+++ packages/form-utils/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.1.0](https://github.com/Availity/element/compare/@availity/mui-form-utils@1.0.1...@availity/mui-form-utils@1.1.0) (2025-03-21)
+
+### Dependency Updates
+
+* `mui-menu` updated to version `1.0.1`
+
+### Features
+
+* **mui-textfield, mui-form-utils:** add placeholder support to select ([3a6b1ef](https://github.com/Availity/element/commit/3a6b1ef2e5ceedca632bd1916734e56402371ffb))
+
 ## [1.0.1](https://github.com/Availity/element/compare/@availity/mui-form-utils@1.0.0...@availity/mui-form-utils@1.0.1) (2025-03-07)
 
 ### Dependency Updates