+++ packages/modal/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.0.2](https://github.com/Availity/element/compare/@availity/mui-modal@1.0.1...@availity/mui-modal@1.0.2) (2025-03-21)
+
+### Dependency Updates
+
+* `mui-card` updated to version `1.0.1`
 ## [1.0.1](https://github.com/Availity/element/compare/@availity/mui-modal@1.0.0...@availity/mui-modal@1.0.1) (2025-03-07)
 
 ### Dependency Updates