+++ packages/tabs/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.2.0](https://github.com/Availity/element/compare/@availity/mui-tabs@1.1.0...@availity/mui-tabs@1.2.0) (2025-03-24)
+
+
+### Features
+
+* **mui-tabs:** add icon and examples ([af6172d](https://github.com/Availity/element/commit/af6172d25882c03b543fff108424a6875027a5a0))
+
 ## [1.1.0](https://github.com/Availity/element/compare/@availity/mui-tabs@1.0.1...@availity/mui-tabs@1.1.0) (2025-03-13)